backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.1
marshmallow==3.21.3
openai==1.30.1
pytest==8.2.2
pytest-flask==1.3.0
